% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_is_straightforward.R
\name{compare_status}
\alias{compare_status}
\title{Compare status between a ViewFullTable and a fgeo table.}
\usage{
compare_status(path_view_csv, path_stem_dir)
}
\arguments{
\item{path_view_csv}{A file, e.g. "ViewFullTable.csv".}

\item{path_stem_dir}{A directory, e.g. "stem"}
}
\value{
A dataframe.
}
\description{
Compare status between a ViewFullTable and a fgeo table.
}
