% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.validate-package.R
\docType{package}
\name{fgeo.validate-package}
\alias{fgeo.validate}
\alias{fgeo.validate-package}
\title{fgeo.validate: Validate ForestGEO Tables}
\description{
Validate ForestGEO tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.validate}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.validate/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} (0000-0002-1986-7988) [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO (https://forestgeo.si.edu/) [copyright holder]
}

}
\keyword{internal}
