% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_is_expected.R
\name{status_is_expected}
\alias{status_is_expected}
\title{Is \code{fgeo$status} the expected value based on \code{view$Status}?}
\usage{
status_is_expected(data)
}
\arguments{
\item{data}{Dataframe, the output of \code{compare_status()}.}
}
\value{
Logical.
}
\description{
Is \code{fgeo$status} the expected value based on \code{view$Status}?
}
