% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_is_straightforward.R
\name{status_is_straightforward}
\alias{status_is_straightforward}
\title{Is \code{fgeo$status} a straightforward abbreviation of \code{view$Status}?}
\usage{
status_is_straightforward(data)
}
\arguments{
\item{data}{Dataframe, the output of \code{compare_status()}.}
}
\value{
Logical.
}
\description{
The alternative is a case when the status in a census changed based on what
the status is in a later census.
}
