% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_status_equals_stem_dfstatus.R
\name{view_status_equals_stem_dfstatus}
\alias{view_status_equals_stem_dfstatus}
\title{Is the variable Status of a ViewFullTable equal to DFstatus of a stem table?}
\usage{
view_status_equals_stem_dfstatus(path_view_csv, path_stem_dir)
}
\arguments{
\item{path_view_csv}{A file, e.g. "ViewFullTable.csv".}

\item{path_stem_dir}{A directory, e.g. "stem"}
}
\value{
Logical.
}
\description{
Check that \code{DFstatus} is exactly the same as the \code{Status} variable in
ViewFullTable.
}
\examples{
\dontrun{
path_stem_dir <- test_path("private/fgeo/luquillo/stem")
path_view_csv <-
  test_path("private/View/luquillo/ViewFullTable_luquillo.csv")
view_status_equals_fgeo_dfstatus(path_view_csv, path_stem_dir)
}
}
